requirements.txt
# Core dependencies for DeepSeek OCR project
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Web framework and API
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# OCR and image processing
easyocr>=1.7.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# DeepSeek API (if using API instead of local model)
openai>=1.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional: for better performance
accelerate>=0.21.0
bitsandbytes>=0.41.0

# File handling
python-magic>=0.4.27
pathlib>=1.0.1

# Configuration management
pyyaml>=6.0.1
configparser>=5.3.0